Detect port properly in vnc_auto.html when used on port 80 or 443.
Browsers (such as Chrome) don't report port numbers in window.location.port when
used on standard ports such as 80 and 443. This causes vnc_auto.html to not
automatically find the port as it should.

This simple change checks if window.location.port is blank, and sets port as
appropriate from the name of the protocol in use.